requirements.txt
asgiref==3.4.1
black==22.1.0
click==8.0.3
Django==3.1.12
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Faker==11.3.0
gunicorn==20.1.0
loguru==0.6.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tomli==2.0.0
typing-extensions==3.10.0.0
